Added support for actions.
This makes it possible to fetch and execute the actions an accessible-object
offers. I also added a context-menu to the accessible-tree in our manual-app
to disaplay and execute actions offered by accessible-objects. And guess
what, it just works (TM) :-) Additional work is needed to figure out why
atspi offers 4 times the same "Activate accessible object" action...