BIC - Version 1: Make AccessibleObject a QObject
This is an experiment for a potential version 2 of the library. The
current lib makes all dbus calls syncronous which is painfully slow
sometimes. By using QObject and properties with notifications it may be
possible to make it asyncronous but still easy to use.

As a side benefit this makes scripting much easier.