Remove Qt 4 code
Version 0.2.0 will be the last release to support Qt 4, there is no app
that is suddenly come out of the grave asking for this library.

The state of Simon is that it needs to be ported to KF5 in any case and
can use the old version of this library. Neither is in very good shape,
let's rather move forward and make maintenance easier.

Reviewers: millansoto, aacid

Subscribers: fux

Differential Revision: https://phabricator.kde.org/D6900